backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
websockets==12.0
scapy==2.5.0
pydantic==2.6.1
python-whois==0.8.0
validators==0.22.0
requests==2.31.0
python-dotenv==1.0.0
